计算胜率的代码移动到源码目录
